app/producer_helpers/requirements.txt
boto3==1.36.26
kafka-python==2.0.2
requests==2.32.3
